Add code to get commit data from the master branch using current methods
