+++ Data-Collection/dataCollection.py
@@ -4,6 +4,7 @@ from libs.databaseConnector import DatabaseConnector
 from repository import Repository
 from languages import Languages
 from branches import Branches
+from commits import Commits
 
 
 class DataCollection:
@@ -31,18 +32,21 @@ class DataCollection:
             "CREATE TABLE Branches (ID INTEGER, Name TEXT, SHA TEXT, PRIMARY KEY(ID))"
         )
 
+        commitsSQL = "CREATE TABLE Commits (ID INTEGER, SHA TEXT, Author TEXT, Commit_Date TEXT, Tree_SHA TEXT, Comment_Count INTEGER, PRIMARY KEY(ID));"
+
         languagesSQL = "CREATE TABLE Languages (ID INTEGER, Language TEXT, Bytes_of_Code INTEGER, PRIMARY KEY(ID))"
 
         repositorySQL = "CREATE TABLE Repository (ID INTEGER, Name TEXT, Owner TEXT, Private TEXT, Fork TEXT, Created_At TEXT, Updated_At TEXT, Pushed_At TEXT, Size INTEGER, Forks INTEGER, Open_Issues INTEGER, PRIMARY KEY(ID))"
 
         self.dbConnector.executeSQL(sql=branchesSQL, commit=True)
+        self.dbConnector.executeSQL(sql=commitsSQL, commit=True)
         self.dbConnector.executeSQL(sql=languagesSQL, commit=True)
         self.dbConnector.executeSQL(sql=repositorySQL, commit=True)
 
     def startDataCollection(self) -> None:
         def _collectData(collector) -> None:
             while True:
-                print("...\tDownloading information")
+                # print("...\tDownloading information")
                 data = collector.getData()
                 collector.insertData(dataset=data[0])
                 if not collector.iterateNext(data[1]):
@@ -57,6 +61,14 @@ class DataCollection:
             repository=self.repository,
             username=self.username,
             url="https://api.github.com/repos/{}/{}/branches?per_page=100&page={}",
+        )
+
+        commitsCollector = Commits(
+            dbConnection=self.dbConnector,
+            oauthToken=self.token,
+            repository=self.repository,
+            username=self.username,
+            url="https://api.github.com/repos/{}/{}/commits?per_page=100&page={}",
         )
 
         languageCollector = Languages(
@@ -78,6 +90,7 @@ class DataCollection:
         _collectData(languageCollector)  # One request only
         _collectData(repositoryCollector)  # One request only
         _collectData(branchCollector)  # Estimated < 10 requests
+        _collectData(commitsCollector)  # Estimated to have the most requests
 
 
 if __name__ == "__main__":
