+++ Data-Collection/libs/collector.py
@@ -27,6 +27,11 @@ class Collector:
         return [response.json(), response]
 
     def iterateNext(self, responseHeaders: Response) -> bool:
+        # print(
+        #     self.githubConnection.parseResponseHeaders(responseHeaders)[
+        #         "X-RateLimit-Remaining"
+        #     ]
+        # )
         if (
             self.githubConnection.parseResponseHeaders(responseHeaders)["Last-Page"]
             == -1
