+++ Data-Collection/repository/notes.txt
@@ -2,8 +2,6 @@ What does this repository contain?
 
 It contains the Data Collection portion of this project that:
 
-The branches in the repository (https://docs.github.com/en/free-pro-team@latest/rest/reference/repos#list-branches)
-
 The commits under a specific branch (https://docs.github.com/en/free-pro-team@latest/rest/reference/repos#list-commits)
 
 The forks of the repository (https://docs.github.com/en/free-pro-team@latest/rest/reference/repos#list-forks)
